+++ NEWS.md
@@ -1,3 +1,68 @@
+# vip 0.1.3.9000
+
+## Enhancements
+
+* Added support for SHAP-based feature importance which makes use of the recent [fastshap](https://cran.r-project.org/package=fastshap) package on CRAN. To use, simply call `vi_shap()` or `vi()` and specify `method = "shap"` [(#87)](https://github.com/koalaverse/vip/issues/87).
+
+* Added support for the [parsnip](https://cran.r-project.org/package=parsnip), [mlr](https://cran.r-project.org/package=mlr), and [mlr3](https://cran.r-project.org/package=mlr3) packages [(#94)](https://github.com/koalaverse/vip/issues/94).
+
+* Added support for `"mvr"` objects from the [pls](https://cran.r-project.org/package=pls) package (currently just calls `caret::varImp()`) [(#35)](https://github.com/koalaverse/vip/issues/35).
+
+* The `"lm"` method for `vi_model()` gained a new `type` argument that allows users
+to use either (1) the raw coefficients if the features were properly standardized (`type = "raw"`), or (2) the absolute value of the corresponding *t*- or *z*-statistic (`type = "stat"`, the default) [(#77)](https://github.com/koalaverse/vip/issues/77).
+
+* New function `gen_friedman()` for simulating data from the Friedman 1 benchmark problem; see `?vip::gen_friedman` for details.
+
+
+## User-visible changes
+
+* The `vi_pdp()` and `vi_ice()` functions have been deprecated and merged into a single new function called `vi_firm()`. Consequently, setting `method = "pdp"` and `method = "ice"` has also been deprecated; use `method = "firm"` instead.
+
+* The `metric` and `pred_wrapper` arguments to `vi_permute()` are no longer optional.
+
+* The `vip()` function gained a new argument, `geom`, for specifying which type of plot to construct. Current options are `geom = "col"` (the default), `geom = "point"`, `geom = "boxplot"`, or `geom = "violin"` (the latter two only work for permutation-based importance with `nsim > 1`) [(#79)](https://github.com/koalaverse/vip/issues/79). Consequently, the `bar` argument has been removed.
+
+* The `vip()` function gained two new arguments for specifying aesthetics: `mapping` and `aesthetics` (for fixed aesthetics like `color = "red"`). Consequently, the arguments `color`, `fill`, etc. have been removed [(#80)](https://github.com/koalaverse/vip/issues/80).
+
+An example illustrating the above two changes is given below:
+```r
+# Load required packages
+library(ggplot2)  # for `aes_string()` function
+
+# Load the sample data
+data(mtcars)
+
+# Fit a linear regression model
+model <- lm(mpg ~ ., data = mtcars)
+
+# Construct variable importance plots
+p1 <- vip(model)
+p2 <- vip(model, mapping = aes_string(color = "Sign"))
+p3 <- vip(model, type = "dotplot")
+p4 <- vip(model, type = "dotplot", mapping = aes_string(color = "Variable"),
+          aesthetics = list(size = 3))
+grid.arrange(p1, p2, p3, p4, nrow = 2)
+```
+
+* The `vip()` function gained a new argument, `include_type`, which defaults to `FALSE`. If `TRUE`, the type of variable importance that was computed is included in the appropriate axis label. Set `include_type = TRUE` to revert to the old behavior.
+
+## Miscellaneous
+
+* Removed dependency on [ModelMetrics](https://cran.r-project.org/package=ModelMetrics) and the built-in family of performance metrics (`metric_*()`) are now documented and exported. See, for example, `?vip::metric_rmse` [(#93)](https://github.com/koalaverse/vip/issues/93).
+
+* Switched to the [tinytest](https://cran.r-project.org/package=tinytest) framework [(#82)](https://github.com/koalaverse/vip/issues/82).
+
+* Minor documentation improvements.
+
+## Bug fixes
+
+* The internal (i.e., not exported) `get_feature_names()` function does a better job with `"nnet"` objects containing factors. It also does a better job at extracting feature names from model objects containing a `"formula"` component.
+
+* `vi_model()` now works correctly for `"glm"` objects with non-Gaussian families (e.g., logistic regression) [(#74)](https://github.com/koalaverse/vip/issues/74).
+
+* Added appropriate **sparklyr** version dependency [(#59)](https://github.com/koalaverse/vip/issues/59).
+
+
 # vip 0.1.3
 
 ## New functions
@@ -19,6 +84,7 @@
 * The `FUN` argument to `vi()`, `vi_pdp()`, and `vi_ice()` has been changed to `var_fun`.
 
 * Only the predicted class probabilities for the reference class are required (as a numeric vector) for binary classification when `metric = "auc"` or `metric = "logloss"`.
+
 
 ## Minor changes